Pull request par failiem

+++ migrations/1702123679-add-absol-magnitude-h-to-ast-table.sql
@@ -1,1 +1,1 @@
-ALTER TABLE ast_daily ADD COLUMN absol_magnitude_h float(90) NOT NULL;
+ALTER TABLE ast_daily ADD COLUMN absol_magnitude_h float(9,2) NOT NULL;
